\name{gen}
\alias{gen}
\alias{mar}
\alias{individual}
\title{
 genotype example data 
}
\description{
GCIM format of DH genotype dataset.
}
\usage{
  data(gen)
}
\details{
Dataset input of file for WangF function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
