\name{genf2}
\alias{genf2}
\alias{f2mar}
\alias{f2individual}
\title{
 genotype example data 
}
\description{
GCIM format of F2 genotype dataset.
}
\usage{
  data(genf2)
}
\details{
Dataset input of file for WenF function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}

