\name{map}
\alias{map}
\alias{marker}
\alias{chr}
\alias{pos}
\title{
 map example data 
}
\description{
GCIM format of DH map dataset.
}
\usage{
  data(map)
}
\details{
Dataset input of file for WangF function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
