\name{mapf2}
\alias{mapf2}
\alias{f2marker}
\alias{f2chr}
\alias{f2pos}
\title{
 map example data 
}
\description{
GCIM format of F2 map dataset.
}
\usage{
  data(mapf2)
}
\details{
Dataset input of file for WenF function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}

