\name{phe}
\alias{phe}
\alias{posi}
\alias{trait1}
\alias{trait2}
\alias{trait3}
\title{
 phenotype example data 
}
\description{
GCIM format of DH phenotype dataset.
}
\usage{
  data(phe)
}
\details{
Dataset input of file for WangF function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
