\name{phef2}
\alias{phef2}
\alias{f2posi}
\alias{f2trait1}
\alias{f2trait2}
\alias{f2trait3}
\title{
 phenotype example data 
}
\description{
GCIM format of F2 phenotype dataset.
}
\usage{
  data(phef2)
}
\details{
Dataset input of file for WenF function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}

